+++ Source/UEProjectIntelligence/Private/Edit/UEPIEditOperationRegistry.cpp
@@ -2,6 +2,7 @@
 
 #include "HAL/PlatformMisc.h"
 #include "Operations/UEPIActorOperations.h"
+#include "Operations/UEPIContentOperations.h"
 #include "Operations/UEPIInputOperations.h"
 #include "Operations/UEPIMaterialOperations.h"
 
@@ -209,6 +210,7 @@ namespace UE::ProjectIntelligence
 			if (Item.Domain == TEXT("actor")) RegisterOperation(MakeUEPIActorOperation(Item));
 			else if (Item.Domain == TEXT("input")) RegisterOperation(MakeUEPIInputOperation(Item));
 			else if (Item.Domain == TEXT("material")) RegisterOperation(MakeUEPIMaterialOperation(Item));
+			else if (Item.Domain == TEXT("content") || Item.Name == TEXT("asset.set_properties")) RegisterOperation(MakeUEPIContentOperation(Item));
 			else RegisterOperation(MakeShared<FUEPIRegisteredBridgeOperation>(Item));
 		}
 	}